+++ assets/javascripts/kitten/components/cards/crowdfunding-card/components/state.js
@@ -5,6 +5,7 @@ import COLORS from '../../../../constants/colors-config'
 import { Text } from '../../../../components/typography/text'
 import { pxToRem } from '../../../../helpers/utils/typography'
 import { KissKissBankBankIcon } from '../../../../components/icons/kisskissbankbank-icon'
+import { ScreenConfig } from '../../../../constants/screen-config'
 
 const COMPONENT_GUTTER = pxToRem(10)
 
@@ -17,6 +18,10 @@ const StyledState = styled.div`
   background-color: ${COLORS.background1};
   padding: ${COMPONENT_GUTTER};
   line-height: 1;
+
+  @media (max-width: ${pxToRem(ScreenConfig.S.max)}) {
+    display: none;
+  }
 `
 
 const StyledWidgetContainer = styled.div`